feat(Add option to programmatically pass creds)
This change allows the setting of a dict with credentials to use
for AWS. skew.awsclient.aws_creds = {'access_key': key, 'secret_key':
secret, 'token': token}. It also requires the setting of a config
placeholder. skew.config._config = {'account' {'012345689': None}}

When set, these will bypass looking for config files to pull these
settings from. This allows for programmatic use of the library
without the need for manipulating files on the local filesytem.