Pass AWS creds to `scan` for thread safety
It's useful to be able to programmatically specify AWS credentials, but
the way that currently works is not thread safe. Passing creds
explicitly through the entire "scan" workflow should fix that.

Side note: Looking at the _client_cache in skew.awsclient, I think
another problem may occur. The thread safety here isn't such a big issue
(since the cache is keyed by the account ID), but I suspect that a
long-running process--such as a web server process--will run into
problems if the tokens are temporary. I suspect this because I've seen
errors like "The security token included in the request is expired" in a
production system, and basically the problem doesn't go away until the
process/container is restarted.